adds delete confirm and updates button styling

+++ app_client/exam/exam.controller.js
@@ -120,15 +120,17 @@ function examController($location, $http, authentication, Flash) {
 
   vm.deleteExam = function (exam) {
     console.log(exam)
-    $http({
-      method: 'DELETE',
-      url: '/api/exams/'+ exam._id
-    }).then(function onSuccess(response) {
-      var index = vm.examsList.indexOf(exam);
-      vm.examsList.splice(index,1)
-    }, function errorCallback(response) {
-      console.log('DELETE error ', response);
-    });
+    if (confirm('Are you sure you want to delete this?')) {
+      $http({
+        method: 'DELETE',
+        url: '/api/exams/'+ exam._id
+      }).then(function onSuccess(response) {
+        var index = vm.examsList.indexOf(exam);
+        vm.examsList.splice(index,1)
+      }, function errorCallback(response) {
+        console.log('DELETE error ', response);
+      });
+    }
   }
 
   vm.errorAlert = function () {
